Fix: Teams list 500 — cast count(*) to int for TeamSummary
Dapper's constructor-materialization path is strict about types:
count(*) returns bigint, TeamSummary.MemberCount is int. Cast ::int in
the query (same fix already applied elsewhere). Added a teams-list
assertion to the identity contract test.

+++ api/GitSrv.Api/Identity/TeamService.cs
@@ -34,7 +34,7 @@ public sealed class TeamService(Db db)
         return (await conn.QueryAsync<TeamSummary>(
             """
             SELECT t.id, t.slug, t.name,
-                   (SELECT count(*) FROM team_members tm WHERE tm.team_id = t.id) AS MemberCount
+                   (SELECT count(*) FROM team_members tm WHERE tm.team_id = t.id)::int AS MemberCount
             FROM teams t WHERE t.org_id = @orgId ORDER BY t.name
             """, new { orgId })).ToList();
     }
